Implement FollowBehaviour class
This commit introduces the FollowBehaviour class, which implements the Behaviour interface. The FollowBehaviour is responsible for determining the appropriate action for an Actor to follow another target Actor on the game map. The AlienBug class is also refactored to include FollowBehaviour. The Application class is modified for testing.

FollowBehaviour
- Constructor that takes the target Actor to follow
- Implemented the getAction() method, which calculates the shortest path towards the target Actor and returns a MoveActorAction to move in that direction
- Implemented a distance() method that computes the Manhattan distance between two locations
- Referred to mars demo

AlienBug
- Introduced new playerToFollow variable
- Added FollowBehaviour with priority 0

Application
- Added craters spawning AlienBugs for testing